refactor: split CoinJoin interfaces out of wallet into their own files
