+++ src/interfaces/wallet.h
@@ -8,6 +8,7 @@
 #include <amount.h>                    // For CAmount
 #include <fs.h>                        // For fs::path
 #include <interfaces/chain.h>          // For ChainClient
+#include <interfaces/coinjoin.h>       // For CoinJoin::*
 #include <pubkey.h>                    // For CKeyID and CScriptID (definitions needed in CTxDestination instantiation)
 #include <script/standard.h>           // For CTxDestination
 #include <support/allocators/secure.h> // For SecureString
@@ -26,7 +27,6 @@
 
 class CCoinControl;
 class CFeeRate;
-class CoinJoinWalletManager;
 class CKey;
 class CWallet;
 class UniValue;
@@ -51,34 +51,6 @@ struct WalletTxStatus;
 
 using WalletOrderForm = std::vector<std::pair<std::string, std::string>>;
 using WalletValueMap = std::map<std::string, std::string>;
-
-namespace CoinJoin {
-//! Interface for the wallet constrained src/coinjoin part of a dash node (dashd process).
-class Client
-{
-public:
-    virtual ~Client() {}
-    virtual void resetCachedBlocks() = 0;
-    virtual void resetPool() = 0;
-    virtual int getCachedBlocks() = 0;
-    virtual std::string getSessionDenoms() = 0;
-    virtual void setCachedBlocks(int nCachedBlocks) = 0;
-    virtual void disableAutobackups() = 0;
-    virtual bool isMixing() = 0;
-    virtual bool startMixing() = 0;
-    virtual void stopMixing() = 0;
-};
-class Loader
-{
-public:
-    virtual ~Loader() {}
-    //! Add new wallet to CoinJoin client manager
-    virtual void AddWallet(CWallet&) = 0;
-    //! Remove wallet from CoinJoin client manager
-    virtual void RemoveWallet(const std::string&) = 0;
-    virtual std::unique_ptr<CoinJoin::Client> GetClient(const CWallet& wallet) = 0;
-};
-}
 
 //! Interface for accessing a wallet.
 class Wallet
@@ -471,9 +443,6 @@ std::unique_ptr<Wallet> MakeWallet(const std::shared_ptr<CWallet>& wallet);
 //! Return implementation of ChainClient interface for a wallet loader. This
 //! function will be undefined in builds where ENABLE_WALLET is false.
 std::unique_ptr<WalletLoader> MakeWalletLoader(Chain& chain, const std::unique_ptr<CoinJoin::Loader>& coinjoin_loader, ArgsManager& args);
-
-std::unique_ptr<CoinJoin::Client> MakeCoinJoinClient(const CoinJoinWalletManager& walletman, const CWallet& wallet);
-std::unique_ptr<CoinJoin::Loader> MakeCoinJoinLoader(CoinJoinWalletManager& walletman);
 
 } // namespace interfaces
